requirements.txt
python-jenkins==1.1.0
github3.py==1.1.0
GitPython==2.1.11
